refactor: update component imports and styles for better organization and maintainability; introduce new styles for search components
