+++ src/components/weather/search/index.ts
@@ -1,0 +1,4 @@
+export { default as MiniSearchBar } from "./MiniSearchBar";
+export { default as LocationItem } from "./LocationItem";
+export { default as SearchInput } from "./SearchInput";
+export { default as NoResults } from "./NoResults";
